services/auth-service/requirements.txt
# Auth Service 의존성

# FastAPI
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# 데이터베이스
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9
alembic>=1.13.0

# 인증
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6

# 검증
pydantic>=2.5.0
pydantic-settings>=2.1.0
email-validator>=2.1.0

# 환경 변수
python-dotenv>=1.0.0
